+++ bittensor/exceptions/handlers/rollbar/__init__.py
@@ -1,7 +1,6 @@
 import os
 from loguru import logger
 import asyncio
-from bittensor.exceptions.handlers import asyncio_exception_handler
 import rollbar
 from pathlib import Path
 
@@ -37,3 +36,32 @@ def set_runtime_status(status):
     file = Path('/tmp/bt_runstate')
     with file.open("w") as file:
         file.write("%s\n" % status)
+
+
+def asyncio_exception_handler(loop, context):
+    logger.debug("asyncio exception has occured")
+
+    exception: BaseException
+    exception = context['exception']
+
+    exc_info = __get_exc_info(exception)
+    extra_data = __get_extra_data(context)
+
+    rollbar.report_exc_info(exc_info=exc_info, extra_data=extra_data)
+    logger.error(context)
+
+    set_runtime_status("ERR")
+
+def __get_exc_info(exception):
+    _type = type(exception)
+    value = exception
+    tb = exception.__traceback__
+    return _type, value, tb
+
+def __get_extra_data(context):
+    frames = []
+    for elem in context['source_traceback']:
+        frames.append(str(elem))
+    extra_data = "\n".join(frames)
+    return extra_data
+
