+++ bittensor/utils/__init__.py
@@ -1,2 +1,1 @@
-from pathlib import Path
